+++ components/sections/main/main.tsx
@@ -4,6 +4,7 @@ import { ReactElement } from 'react';
 
 // import { Hero } from '../hero/hero';
 import { About } from '../about/about';
+import { Intorduction } from '../introduction/introduction';
 // import { MainSkills } from '../skills/main-skills';
 // import { InProgressSkills } from '../skills/in-progress-skills';
 // import { Experience } from '../experience/experience';
@@ -15,6 +16,7 @@ import * as S from './main.styles';
 export const Main = (): ReactElement => {
   return (
     <S.Container>
+      <Intorduction />
       <About />
       {/* <Hero />
       <MainSkills />
