+++ styles/globals.css
@@ -18,7 +18,7 @@ body {
   min-height: 100vh;
   text-rendering: optimizeSpeed;
   line-height: 1;
-  background-color: grey;
+  background-color: pink;
   width: 100%;
 }
